Remove old create component
