--- src/routes/ReCreate.scss
@@ -1,164 +1,0 @@
-@import '../scss/index';
-
-.re-create {
-  overflow: auto;
-  position: relative;
-  display: block;
-  width: 100vw;
-  min-height: 100vh;
-  background-image: url('../images/bg.jpg');
-  background-repeat: no-repeat;
-  background-position: center top;
-  background-size: cover;
-
-  .create-form {
-    margin: 2rem auto 3rem;
-    padding: 1rem 0 0;
-    width: $screen-lg;
-    color: #fff;
-    background-color: rgba($dark, .85);
-    border: 2px solid $gray;
-
-    @media only screen and (max-width: $screen-md) {
-      width: 90vw;
-    }
-  }
-}
-
-.create-form {
-  .require::after {
-    content: '*';
-    margin-left: .15rem;
-    color: $danger;
-    font-family: $header-font;
-  }
-
-  .create-title {
-    margin-bottom: 2rem;
-    padding: .5rem 0 1rem;
-    text-align: center;
-    font-family: $header-font;
-    color: #fff;
-    border-bottom: 1px solid $gray;
-  }
-
-  .create-thumbnails {
-    margin-top: 1rem;
-    width: 100%;
-
-    img {
-      margin: 0 .5rem;
-      height: 80px;
-      border: 2px solid transparent;
-      opacity: .5;
-      transition: all .2s ease;
-      cursor: pointer;
-
-      &:hover {
-        opacity: 1;
-      }
-
-      &.selected {
-        border: 2px solid $light-warning;
-        opacity: 1;
-      }
-    }
-  }
-
-  .create-file {
-    margin: 1rem 0;
-    text-align: center;
-    cursor: pointer;
-
-    .files-title {
-      overflow: hidden;
-      padding: 2rem;
-      max-width: 365px;
-      width: 365px;
-      font-size: 17px;
-      text-overflow: ellipsis;
-      background-color: rgba($dark, .85);
-      color: #fff;
-      letter-spacing: 1px;
-      border: 1px solid $gray;
-      border-radius: .5rem;
-      white-space: nowrap;
-      transition: all .2s ease-in-out;
-
-      &:hover {
-        background-color: rgba($gray, .55);
-      }
-
-      + p {
-        height: 20px;
-        color: $warning;
-        text-align: center;
-
-        &::before {
-          content: '【 ';
-        }
-
-        &::after {
-          content: ' 】';
-        }
-      }
-    }
-
-    .files-message {
-      display: block;
-      margin-top: .5rem;
-      color: $danger;
-    }
-  }
-
-  .create-item {
-    margin-bottom: 1.75rem;
-
-    > h2 {
-      font-size: 20px;
-      width: 100%;
-      margin-bottom: .5rem;
-      font-family: $global-font;
-    }
-
-    > input {
-      padding: .65rem .5rem;
-      font-size: 16px;
-      font-family: $input-font;
-    }
-  }
-
-  .create-button {
-    justify-content: center;
-    border-top: 1px solid $gray;
-
-    button {
-      margin-top: 1.5rem;
-      width: 200px;
-      height: 40px;
-      font-family: $global-font;
-
-      &:first-child {
-        margin-right: 1rem;
-      }
-    }
-  }
-
-  /* overwrite */
-  .p-grid {
-    margin: 0 0 1.5rem;
-    padding: 0 1.5rem;
-    max-width: 100%;
-  }
-
-  .p-editor-container {
-    width: 100%;
-
-    /* Edit content basic style */
-    .p-editor-content.ql-container,
-    .ql-editor {
-      font-size: 17px;
-      line-height: 1.7;
-    }
-  }
-}
